Move MonacoEditor to a separate component
