+++ README.md
@@ -2,4 +2,8 @@
 
 This library configures a [monaco editor](https://github.com/microsoft/monaco-editor) to use [@cucumber/language-service](https://github.com/cucumber/language-service#readme)
 
-[Live demo](https://cucumber.github.io/monaco/)
+[Live demo](https://cucumber.github.io/monaco/).
+
+## Usage
+
+The library can be used with vanilla DOM as well as React. See [try/index.tsx](try/index.tsx) (the source for the live demo) for details.
